Settings: Remove circular dependency on CPowerManager
