+++ xbmc/Application.cpp
@@ -74,8 +74,9 @@
 #include "playlists/PlayList.h"
 #include "profiles/ProfilesManager.h"
 #include "windowing/WinSystem.h"
-#include "powermanagement/PowerManager.h"
 #include "powermanagement/DPMSSupport.h"
+#include "powermanagement/PowerManager.h"
+#include "powermanagement/PowerTypes.h"
 #include "settings/Settings.h"
 #include "settings/AdvancedSettings.h"
 #include "settings/DisplaySettings.h"
