+++ xbmc/powermanagement/CMakeLists.txt
@@ -4,7 +4,8 @@ set(SOURCES DPMSSupport.cpp
 
 set(HEADERS DPMSSupport.h
             IPowerSyscall.h
-            PowerManager.h)
+            PowerManager.h
+            PowerTypes.h)
 
 if(CORE_SYSTEM_NAME MATCHES windows)
   list(APPEND HEADERS WinIdleTimer.h)
